docs(utils): move utils for React Static into separate directory
