--- docs/src/staticUtils/requireInfo.js
@@ -1,3 +1,0 @@
-const requireInfo = displayName => require(`../componentInfo/${displayName}.info.json`)
-
-export default requireInfo
