+++ docs/static/utils/getComponentGroupInfo.js
