+++ docs/static/utils/getExampleSources.js
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import path from 'path'
 
-const filename = path.resolve(__dirname, '..', 'exampleSources.json')
+const filename = path.resolve(__dirname, '../../src/exampleSources.json')
 
 const getExampleSources = () => JSON.parse(fs.readFileSync(filename))
