+++ docs/static/utils/index.js
@@ -3,4 +3,5 @@ export getComponentMenu from './getComponentMenu'
 export getExampleSources from './getExampleSources'
 export getInfoForSeeTags from './getInfoForSeeTags'
 export getLayoutPaths from './getLayoutPaths'
+export getPagesPaths from './getPagesPaths'
 export getSidebarSections from './getSidebarSections'
