+++ docs/static/utils/requireInfo.js
@@ -1,0 +1,3 @@
+const requireInfo = displayName => require(`../../src/componentInfo/${displayName}.info.json`)
+
+export default requireInfo
